Object item refactoring (splitted into several files)
